requirements.txt
kagglehub==0.3.13
pandas==2.3.3
numpy==2.3.4
matplotlib==3.10.7
sqlalchemy==2.0.44
python-dotenv==1.2.1
